src/main/java/com/aalvarezg/ms_usuarios/infrastructure/ouput/jpa/adapter/BCryptIPasswordEncoderAdapter.java
package com.aalvarezg.ms_usuarios.infrastructure.ouput.jpa.adapter;

import com.aalvarezg.ms_usuarios.domain.spi.IPasswordEncoderPort;
import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
import org.springframework.stereotype.Component;

@Component
public class BCryptIPasswordEncoderAdapter implements IPasswordEncoderPort {

    private final BCryptPasswordEncoder encoder = new BCryptPasswordEncoder();
    @Override
    public String encode(String rawPassword) {
        return encoder.encode(rawPassword);
    }

    @Override
    public boolean matches(String rawPassword, String encodedPassword) {
        return encoder.matches(rawPassword, encodedPassword);
    }
}
